+++ app/src/App.tsx
@@ -11,12 +11,12 @@ import {
   RadioGroup,
   RadioGroupItem,
   Textarea,
+  Toggle,
   Tooltip,
 } from '@moondreamsdev/dreamer-ui/components';
 import { useActionModal, useToast } from '@moondreamsdev/dreamer-ui/hooks';
 import { ActionModalProvider, ToastProvider } from '@moondreamsdev/dreamer-ui/providers';
 import { useState } from 'react';
-import { Toggle } from './toggle';
 
 const TestComponent = ({ index }: { index: number }) => {
   return (
