--- app/src/toggle/index.ts
@@ -1,2 +1,0 @@
-export { default as Toggle } from './Toggle';
-export type { ToggleProps } from './Toggle';
