+++ lib/src/components/toggle/index.ts
@@ -1,0 +1,2 @@
+export { default as Toggle } from './Toggle';
+export type { ToggleProps } from './Toggle';
